procfs: Fix asprintf format
Last_processor is an int in the proc info, and there aren't
really plans for machines with more than 4 billion processors.

* procfs/process.c (process_file_gc_stat): Make last_processor unsigned
instead of long unsigned.

+++ procfs/process.c
@@ -231,7 +231,7 @@ process_file_gc_stat (struct proc_stat *ps, char **contents)
   process_t p;
   error_t err = proc_pid2proc (ps->context->server, ps->pid, &p);
 
-  long unsigned last_processor;
+  unsigned last_processor;
 
 #ifdef HAVE_STRUCT_THREAD_SCHED_INFO_LAST_PROCESSOR
   last_processor = thsi->last_processor;
